implemented a progress circle component
a usable component. fiddled with Victory Native Lib for data visualization. Their docs are pretty vague so I don't 100% understand the code but it works, you can import and pass props (ctrl + space in vsc to se what props are available, but for more details on what props mean you have to look intp progressCircle.js - source code for this component) to customize.