Fixed bugs with date time picker
Namely the one where it would crash on cancel, open up the dialog twice, and sometimes leave a white screen instead of opening the picker. Furthermore, I got rid of modal in this component as I believe the datepicker is a modal itself, so no need to make room for error and bugs with even more layered modals. Now it uses the state directly, as shown in the example at: https://github.com/react-native-datetimepicker/datetimepicker#basic-usage-with-state