Colors, styles, icons cleanup
Centralized styles and colors here and there, but I don't know if that's enough. It would be nice if you could look into the code and see if some style could be centralized. I feel like these are hard decisions to make. And from now on, we try to only use centralized colors for consistency